requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
aws-requests-auth==0.4.3
boto3==1.38.42
botocore==1.38.42
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
dill==0.4.0
docstring_parser==0.16
fastapi==0.115.13
fonttools==4.58.4
h11==0.16.0
halo==0.0.31
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
importlib_metadata==8.7.0
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
log-symbols==0.0.14
markdown-it-py==3.0.0
matplotlib==3.10.3
mcp==1.13.1
mdurl==0.1.2
mpmath==1.3.0
numpy==2.3.1
ollama==0.5.1
opentelemetry-api==1.34.1
opentelemetry-exporter-zipkin-json==1.35.0
opentelemetry-exporter-zipkin-proto-http==1.35.0
opentelemetry-instrumentation==0.55b1
opentelemetry-instrumentation-threading==0.55b1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
packaging==25.0
pandas==2.2.3
pillow==11.2.1
openpyxl==3.1.5
xlrd==2.0.1
prompt_toolkit==3.0.51
protobuf==3.20.3
psycopg2-binary==2.9.10
mysql-connector-python==9.3.0
pydantic==2.11.7
pydantic-settings==2.10.0
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pywin32==311; sys_platform == "win32"
referencing==0.36.2
requests==2.32.4
rich==14.0.0
rpds-py==0.27.0
s3transfer==0.13.0
six==1.17.0
slack_bolt==1.23.0
slack_sdk==3.35.0
sniffio==1.3.1
spinners==0.0.24
sse-starlette==2.3.6
starlette==0.46.2
strands-agents==1.17.0
sympy==1.14.0
tenacity==9.1.2
termcolor==3.1.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.34.3
watchdog==6.0.0
wcwidth==0.2.13
wrapt==1.17.3
zipp==3.23.0
sshtunnel==0.4.0
paramiko==2.11.0
pytz>=2023.3,<2024.0
mysql-connector==2.2.9
git+https://{GITHUB_TOKEN}@github.com/{GITHUB_REPO}.git@{GITHUB_BRANCH}